+++ ServiceManager/src/smgl/GLImportBatchesAction.java
@@ -61,6 +61,8 @@ public class GLImportBatchesAction extends HttpServlet{
 	private static final int FIELD_LINE_CREDITAMT = 17;
 	private static final int FIELD_LINE_SOURCETYPE = 18;
 	
+	private static String IMPORT_FILE_PREFIX = "GLIMPORT_";
+	
 	//Fields for older 'ACCPAC STYLE' import files:
 	private static final int ACCPAC_FORMAT_FIELD_RECORD_TYPE = 0;
 
@@ -151,22 +153,7 @@ public class GLImportBatchesAction extends HttpServlet{
 		
 		return;
 	}
-	private void createTempImportFileFolder(String sTempFileFolder) throws Exception{
-	    File dir = new File(sTempFileFolder);
-	    if (dir.exists()) {
-	      return;
-	    }
-	    
-	    //Need to create the path:
-	    try{
-	        // Create one directory
-	        if (!new File(sTempFileFolder).mkdir()) {
-	        	throw new Exception("Error [1557852631] creating temp upload folder.");
-	        }    
-        }catch (Exception e){//Catch exception if any
-        	throw new Exception("Error [1557852632] creating temp upload folder - " + e.getMessage() + ".");
-	    }
-	}
+
 	private void processRequest(
 			HttpSession ses, 
 			HttpServletRequest req,
@@ -177,20 +164,15 @@ public class GLImportBatchesAction extends HttpServlet{
 			HashMap<String, String> mv
 			) throws Exception{
 
-    	String sTempFilePath = SMUtilities.getAbsoluteRootPath(req, getServletContext())
-			+ System.getProperty("file.separator")
-			+ "iccountuploads"
-		;
-
-    	//If the folder has not been created, create it now:
-		try {
-			createTempImportFileFolder(sTempFilePath);
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());
+		String sTempFilePath = SMUtilities.getAbsoluteSMTempPath(req, getServletContext());
+		//Strip off the trailing file separator character, if it has one:
+		if (sTempFilePath.endsWith(System.getProperty("file.separator"))){
+			sTempFilePath = sTempFilePath.substring(0, sTempFilePath.length() - 1);
 		}
+
 		//First, remove any temporary files:
 		try {
-			deleteCurrentTempImportFiles(sTempFilePath);
+			deleteCurrentImportFiles(sTempFilePath);
 		} catch (Exception e1) {
 			throw new Exception("Error [1557852633] deleting temp files before import - " + e1.getMessage());
 		}
@@ -205,11 +187,10 @@ public class GLImportBatchesAction extends HttpServlet{
 			throw new Exception (e2.getMessage());
 		}
 		try {
-			deleteCurrentTempImportFiles(sTempFilePath);
+			deleteCurrentImportFiles(sTempFilePath);
 		} catch (Exception e3) {
 			throw new Exception("Error [1557852635] deleting temp files after import - " + e3.getMessage());
 		}
-		
 	}
 	@SuppressWarnings("unchecked")
 	private void writeFileAndProcess(
@@ -250,7 +231,7 @@ public class GLImportBatchesAction extends HttpServlet{
 			throw new Exception("Error [1557852636] on upload.parseRequest: " + e1.getMessage());
 		}
 		Iterator<FileItem> iter = fileItems.iterator();
-		String fileName = "GLIMPORT_" + clsDateAndTimeConversions.now("yyyyMMdd_HHmmss") + ".csv";
+		String fileName = IMPORT_FILE_PREFIX + clsDateAndTimeConversions.now("yyyyMMdd_HHmmss") + ".csv";
 		while (iter.hasNext()) {
 		    FileItem item = (FileItem) iter.next();
 		    if (item.isFormField()) {
@@ -691,24 +672,28 @@ public class GLImportBatchesAction extends HttpServlet{
 		return;
 	}
 
-	private void deleteCurrentTempImportFiles(String sTempImportFilePath) throws Exception{
-		
-	    File dir = new File(sTempImportFilePath);
-	    if (!dir.exists()) {
-	    	throw new Exception("Error [1557852657] - directory " + sTempImportFilePath + " already exists.");
-	    }
-	    String[] info = dir.list();
-	    for (int i = 0; i < info.length; i++) {
-	      File n = new File(sTempImportFilePath + System.getProperty("file.separator") + info[i]);
-	      if (!n.isFile()) { // skip ., .., other directories, etc.
-	        continue;
-	      }
-	      if (!n.delete()){
-	    	  throw new Exception("Error [1557852658] - error deleting " 
-	    		+ sTempImportFilePath + System.getProperty("file.separator") + info[i] + ".");
-	      }
-	    }
+	private void deleteCurrentImportFiles(String sTempFilePath) throws Exception{
 
+		File dir = new File(sTempFilePath);
+		if (!dir.exists()) {
+			//Nothing to do in this case....
+			return;
+		}
+
+		String[] info = dir.list();
+
+		for (int i = 0; i < info.length; i++) {
+			File n = new File(sTempFilePath + info[i]);
+			if (!n.isFile()) { // skip ., .., other directories, etc.
+				continue;
+			}
+			if (info[i].startsWith(IMPORT_FILE_PREFIX)){
+				if (!n.delete()){
+					throw new Exception("Error [20201195104] " + "Unable to delete " + sTempFilePath + info[i]);
+				}
+			}
+		}
+		return;
 	}
 	private void validateFile(String sFilePath, String sFileName, boolean bFileIncludesHeaderRow, Connection conn) throws Exception{
